refactor: clean up and rename randomQuotesComponent.js to RandomQuote.jsx
