+++ src/App.js
@@ -1,7 +1,7 @@
 import { Provider } from 'react-redux';
 import React, { Component } from 'react';
 
-import Quotes from './components/randomQuotesComponent';
+import Quotes from './components/RandomQuote';
 import { ConfigureStore } from './redux/store';
 
 import './App.css';
